requirements.txt
Flask==3.0.2
flask-cors==4.0.0
python-docx==1.1.0
PyMuPDF==1.23.26
Werkzeug==3.0.1
gunicorn==21.2.0
python-magic==0.4.27
python-magic-bin==0.4.14
requests==2.31.0 